Add ProductsController and update server.csproj
The `server.csproj` file has been updated to remove the `<ItemGroup>` that included the "Controllers" folder.

A new `ProductsController` class has been added in the `ProductsController.cs` file. This class is set up as an API controller with the appropriate route and attributes for handling HTTP requests in an ASP.NET Core application.